requirements.txt
google-generativeai
selenium
beautifulsoup4
webdriver-manager
python-dotenv
lxml
google-genai